Migrate to command line execution of algorithms.
Remove html pages.  Add command line logic.
