--- Problem-01/problem-01.html
@@ -1,8 +1,0 @@
-<html>
-<head>
-    <script src="problem-01-multiples.js"></script>
-</head>
-    <body>
-        Problem 1.
-    </body>
-</html>
