--- Problem-02/problem-02-recursive.js
@@ -1,36 +1,0 @@
-/* Project Euler - Problem 2
-ddanaherJS
-
-Each new term in the Fibonacci sequence is generated by adding the previous two terms.
-By starting with 1 and 2, the first 10 terms will be:
-
-1, 2, 3, 5, 8, 13, 21, 34, 55, 89, ...
-
-By considering the terms in the Fibonacci sequence whose values do not exceed four million,
-find the sum of the even-valued terms.
-
-https://projecteuler.net/problem=2
-*/
-
-var startTime = new Date().getMilliseconds();
-var endTime;
-var nextElement = 0;
-var limit = 4000000;
-var sumOfEvens = 2;
-
-console.log('Problem 2: The (recursive) solution [' + fibonacci(1, 2) + '] was found in [' + (endTime - startTime) + '] ms');
-
-function fibonacci(lesser, greater) {
-    nextElement = lesser + greater;
-
-    if (nextElement <= limit) {
-
-        if (nextElement % 2 === 0) {
-            sumOfEvens += nextElement;
-        }
-
-        fibonacci(greater, nextElement);
-    }
-    endTime = new Date().getMilliseconds();
-    return sumOfEvens;
-}
