--- Problem-02/problem-02.html
@@ -1,9 +1,0 @@
-<html>
-<head>
-    <script src="problem-02-even-fibonacci.js"></script>
-    <script src="problem-02-recursive.js"></script>
-</head>
-    <body>
-        Problem 2.
-    </body>
-</html>
