--- Problem-03/problem-03.html
@@ -1,8 +1,0 @@
-<html>
-<head>
-    <script src="problem-03-largest-prime.js"></script>
-</head>
-    <body>
-        Problem 3.
-    </body>
-</html>
