--- Problem-04/problem-04.html
@@ -1,8 +1,0 @@
-<html>
-<head>
-    <script src="problem-04-palindrome.js"></script>
-</head>
-    <body>
-        Problem 4.
-    </body>
-</html>
