--- Problem-05/problem-05.html
@@ -1,9 +1,0 @@
-<html>
-<head>
-    <script src="problem-05-smallest-multiple.js"></script>
-    <script src="problem-05-smallest-multiple-refactor.js"></script>
-</head>
-    <body>
-        Problem 5.
-    </body>
-</html>
