--- Problem-06/problem-06.html
@@ -1,8 +1,0 @@
-<html>
-<head>
-    <script src="problem-06-sum-squares-difference.js"></script>
-</head>
-    <body>
-        Problem 6.
-    </body>
-</html>
