--- Problem-07/problem-07.html
@@ -1,8 +1,0 @@
-<html>
-<head>
-    <script src="problem-07-10001st-prime.js"></script>
-</head>
-    <body>
-        Problem 7.
-    </body>
-</html>
