--- Problem-08/problem-08.html
@@ -1,8 +1,0 @@
-<html>
-<head>
-    <script src="problem-08-largest-product.js"></script>
-</head>
-    <body>
-        Problem 8.
-    </body>
-</html>
