--- Problem-09/problem-09.html
@@ -1,8 +1,0 @@
-<html>
-<head>
-    <script src="problem-09-pythagorean-triple.js"></script>
-</head>
-    <body>
-        Problem 9.
-    </body>
-</html>
