+++ Problems/problem-02-even-fibonacci.js
@@ -1,0 +1,36 @@
+this.description = 'Project Euler - Problem 2'
+    +'\nEach new term in the Fibonacci sequence is generated by adding the '
+    +'\nprevious two terms. By starting with 1 and 2, the first 10 terms will be:'
+    +'\n\n1, 2, 3, 5, 8, 13, 21, 34, 55, 89, ...'
+    +'\n\nBy considering the terms in the Fibonacci sequence whose values '
+    +'\ndo not exceed four million, find the sum of the even-valued terms.'
+    +'\nhttps://projecteuler.net/problem=2';
+
+this.getSolution = function() {
+    return evenFibonacci(4000000);
+}
+
+var lesser = 1;
+var greater = 2;
+var sumOfEvens = 2;
+var currentElement = 0;
+
+function evenFibonacci(upperLimit) {
+
+    while(currentElement <= upperLimit) {
+
+        if (currentElement % 2 === 0) {
+            sumOfEvens += currentElement;
+        }
+
+        currentElement = generateNextFibonacci(lesser, greater);
+        lesser = greater;
+        greater = currentElement;
+    }
+
+    return sumOfEvens;
+}
+
+function generateNextFibonacci(smaller, larger) {
+    return smaller + larger;
+}
