+++ Problems/problem-02-recursive.js
@@ -1,0 +1,30 @@
+this.description = 'Project Euler - Problem 2'
+    +'\nEach new term in the Fibonacci sequence is generated by adding the '
+    +'\nprevious two terms. By starting with 1 and 2, the first 10 terms will be:'
+    +'\n\n1, 2, 3, 5, 8, 13, 21, 34, 55, 89, ...'
+    +'\n\nBy considering the terms in the Fibonacci sequence whose values '
+    +'\ndo not exceed four million, find the sum of the even-valued terms.'
+    +'\nhttps://projecteuler.net/problem=2';
+
+this.getSolution = function() {
+    return fibonacci(1,2);
+}
+
+var nextElement = 0;
+var sumOfEvens = 2;
+var limit = 4000000;
+
+function fibonacci(lesser, greater) {
+    nextElement = lesser + greater;
+
+    if (nextElement <= limit) {
+
+        if (nextElement % 2 === 0) {
+            sumOfEvens += nextElement;
+        }
+
+        fibonacci(greater, nextElement);
+    }
+
+    return sumOfEvens;
+}
