+++ utils/problemList.js
@@ -1,0 +1,16 @@
+var pre = './Problems/';
+
+var problems = {
+    'problem-1': pre + 'problem-01-multiples.js',
+    'problem-2': pre + 'problem-02-recursive.js',
+    'problem-3': pre + 'problem-03-largest-prime.js',
+    'problem-4': pre + 'problem-04-palindrome.js',
+    'problem-5': pre + 'problem-05-smallest-multiple-refactor.js',
+    'problem-6': pre + 'problem-06-sum-squares-difference.js',
+    'problem-7': pre + 'problem-07-10001st-prime.js',
+    'problem-8': pre + 'problem-08-largest-product.js',
+    'problem-9': pre + 'problem-09-pythagorean-triple.js'
+    // 'problem-': pre + 'problem-0',
+}
+
+module.exports = problems;
