--- view.html
@@ -1,17 +1,0 @@
-<html>
-<head>
-</head>
-    <body>
-        <ol>
-            <li><a href="Problem-01/problem-01.html">Problem 1</a></li>
-            <li><a href="Problem-02/problem-02.html">Problem 2</a></li>
-            <li><a href="Problem-03/problem-03.html">Problem 3</a></li>
-            <li><a href="Problem-04/problem-04.html">Problem 4</a></li>
-            <li><a href="Problem-05/problem-05.html">Problem 5</a></li>
-            <li><a href="Problem-06/problem-06.html">Problem 6</a></li>
-            <li><a href="Problem-07/problem-07.html">Problem 7</a></li>
-            <li><a href="Problem-08/problem-08.html">Problem 8</a></li>
-            <li><a href="Problem-09/problem-09.html">Problem 9</a></li>
-        </ol>
-    </body>
-</html>
